% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_backtest_config}
\alias{default_backtest_config}
\title{Default backtest configuration}
\usage{
default_backtest_config()
}
\value{
A list of default parameters.
}
\description{
Returns a list of all default parameters used by the backtesting engine.
}
