% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_visualizations.R
\name{plot_all_visualizations}
\alias{plot_all_visualizations}
\title{Plot all visualizations in a grid}
\usage{
plot_all_visualizations(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object (arranged grid)
}
\description{
Plot all visualizations in a grid
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_all_visualizations(perf)
}
}
