% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asset_position_combined.R
\name{plot_asset_position_combined}
\alias{plot_asset_position_combined}
\title{Combined plot: Equity curve + Position ratio}
\usage{
plot_asset_position_combined(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object (arranged using ggpubr)
}
\description{
Combined plot: Equity curve + Position ratio
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_asset_position_combined(perf)
}
}
