% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_return.R
\name{plot_cumulative_return}
\alias{plot_cumulative_return}
\title{Plot cumulative return area chart}
\usage{
plot_cumulative_return(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Plot cumulative return area chart
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_cumulative_return(perf)
}
}
