% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_return.R
\name{plot_daily_return}
\alias{plot_daily_return}
\title{Plot daily return bar chart}
\usage{
plot_daily_return(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Plot daily return bar chart
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_daily_return(perf)
}
}
