% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drawdown.R
\name{plot_drawdown}
\alias{plot_drawdown}
\title{Plot drawdown area chart}
\usage{
plot_drawdown(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Plot drawdown area chart
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_drawdown(perf)
}
}
