% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_equity_curve.R
\name{plot_equity_curve}
\alias{plot_equity_curve}
\title{Plot equity curve (total asset over time)}
\usage{
plot_equity_curve(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Plot equity curve (total asset over time)
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_equity_curve(perf)
}
}
