% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position_ratio.R
\name{plot_position_ratio}
\alias{plot_position_ratio}
\title{Plot position ratio over time}
\usage{
plot_position_ratio(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Plot position ratio over time
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_position_ratio(perf)
}
}
