% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_return_drawdown_combined.R
\name{plot_return_drawdown_combined}
\alias{plot_return_drawdown_combined}
\title{Combined plot: Cumulative return + Drawdown in one chart}
\usage{
plot_return_drawdown_combined(perf_result)
}
\arguments{
\item{perf_result}{List, output from performance_analysis() function}
}
\value{
ggplot object
}
\description{
Combined plot: Cumulative return + Drawdown in one chart
}
\examples{
\dontrun{
perf <- performance_analysis(result)
plot_return_drawdown_combined(perf)
}
}
