% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_component_stop_loss}
\alias{set_component_stop_loss}
\title{Set component stop-loss parameters}
\usage{
set_component_stop_loss(
  config,
  enable = TRUE,
  type = "fixed",
  fixed_ratio = 0.1,
  trailing_fixed_ratio = 0.1,
  atr_n = 14,
  atr_k = 2,
  vol_n = 20,
  vol_sigma = 2,
  log_vol_n = 20,
  log_vol_sigma = 2
)
}
\arguments{
\item{config}{Configuration list.}

\item{enable}{Logical, enable component stop-loss.}

\item{type}{Stop-loss type.}

\item{fixed_ratio}{Fixed ratio.}

\item{trailing_fixed_ratio}{Trailing fixed ratio.}

\item{atr_n}{ATR lookback.}

\item{atr_k}{ATR multiplier.}

\item{vol_n}{Volatility lookback.}

\item{vol_sigma}{Volatility sigma.}

\item{log_vol_n}{Log volatility lookback.}

\item{log_vol_sigma}{Log volatility sigma.}
}
\value{
Modified configuration list.
}
\description{
Set component stop-loss parameters
}
