% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_component_take_profit}
\alias{set_component_take_profit}
\title{Set component take-profit parameters}
\usage{
set_component_take_profit(
  config,
  enable = TRUE,
  type = "fixed",
  fixed_ratio = 0.1,
  trailing_fixed_ratio = 0.1,
  atr_k = 2,
  vol_sigma = 2,
  log_vol_sigma = 2
)
}
\arguments{
\item{config}{Configuration list.}

\item{enable}{Logical.}

\item{type}{Take-profit type.}

\item{fixed_ratio}{Fixed ratio.}

\item{trailing_fixed_ratio}{Trailing fixed ratio.}

\item{atr_k}{ATR multiplier.}

\item{vol_sigma}{Volatility sigma.}

\item{log_vol_sigma}{Log volatility sigma.}
}
\value{
Modified configuration list.
}
\description{
Set component take-profit parameters
}
