% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_date_range}
\alias{set_date_range}
\title{Set date range}
\usage{
set_date_range(config, start_date = NULL, end_date = NULL)
}
\arguments{
\item{config}{Configuration list.}

\item{start_date}{Start date (Date or character).}

\item{end_date}{End date (Date or character).}
}
\value{
Modified configuration list.
}
\description{
Set date range
}
