% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_parameter}
\alias{set_parameter}
\alias{set_weight_col}
\alias{set_start_date}
\alias{set_end_date}
\alias{set_exec_price_col}
\alias{set_eval_price_col}
\alias{set_init_capital}
\alias{set_lot_size}
\alias{set_fee_rate}
\alias{set_stamp_tax}
\alias{set_slippage_rate}
\alias{set_min_weight}
\alias{set_enable_component_stop_loss}
\alias{set_component_stop_loss_type}
\alias{set_fixed_component_sl_ratio}
\alias{set_trailing_fixed_component_sl_ratio}
\alias{set_atr_n_component}
\alias{set_atr_k_component}
\alias{set_vol_n_component}
\alias{set_vol_sigma_component}
\alias{set_log_vol_n_component}
\alias{set_log_vol_sigma_component}
\alias{set_enable_portfolio_stop_loss}
\alias{set_portfolio_stop_loss_type}
\alias{set_fixed_portfolio_sl_ratio}
\alias{set_trailing_fixed_portfolio_sl_ratio}
\alias{set_atr_n_portfolio}
\alias{set_atr_k_portfolio}
\alias{set_vol_n_portfolio}
\alias{set_vol_sigma_portfolio}
\alias{set_log_vol_n_portfolio}
\alias{set_log_vol_sigma_portfolio}
\alias{set_enable_component_take_profit}
\alias{set_component_take_profit_type}
\alias{set_fixed_component_tp_ratio}
\alias{set_trailing_fixed_component_tp_ratio}
\alias{set_atr_k_component_tp}
\alias{set_vol_sigma_component_tp}
\alias{set_log_vol_sigma_component_tp}
\alias{set_enable_portfolio_take_profit}
\alias{set_portfolio_take_profit_type}
\alias{set_fixed_portfolio_tp_ratio}
\alias{set_trailing_fixed_portfolio_tp_ratio}
\alias{set_atr_k_portfolio_tp}
\alias{set_vol_sigma_portfolio_tp}
\alias{set_log_vol_sigma_portfolio_tp}
\alias{set_single_max_weight}
\alias{set_global_max_hold_pct}
\alias{set_rebalance_mode}
\alias{set_rebalance_cycle}
\alias{set_weight_change_threshold}
\alias{set_skip_suspended}
\alias{set_output_type}
\title{Set backtest parameters}
\usage{
set_weight_col(config, value)

set_start_date(config, value)

set_end_date(config, value)

set_exec_price_col(config, value)

set_eval_price_col(config, value)

set_init_capital(config, value)

set_lot_size(config, value)

set_fee_rate(config, value)

set_stamp_tax(config, value)

set_slippage_rate(config, value)

set_min_weight(config, value)

set_enable_component_stop_loss(config, value)

set_component_stop_loss_type(config, value)

set_fixed_component_sl_ratio(config, value)

set_trailing_fixed_component_sl_ratio(config, value)

set_atr_n_component(config, value)

set_atr_k_component(config, value)

set_vol_n_component(config, value)

set_vol_sigma_component(config, value)

set_log_vol_n_component(config, value)

set_log_vol_sigma_component(config, value)

set_enable_portfolio_stop_loss(config, value)

set_portfolio_stop_loss_type(config, value)

set_fixed_portfolio_sl_ratio(config, value)

set_trailing_fixed_portfolio_sl_ratio(config, value)

set_atr_n_portfolio(config, value)

set_atr_k_portfolio(config, value)

set_vol_n_portfolio(config, value)

set_vol_sigma_portfolio(config, value)

set_log_vol_n_portfolio(config, value)

set_log_vol_sigma_portfolio(config, value)

set_enable_component_take_profit(config, value)

set_component_take_profit_type(config, value)

set_fixed_component_tp_ratio(config, value)

set_trailing_fixed_component_tp_ratio(config, value)

set_atr_k_component_tp(config, value)

set_vol_sigma_component_tp(config, value)

set_log_vol_sigma_component_tp(config, value)

set_enable_portfolio_take_profit(config, value)

set_portfolio_take_profit_type(config, value)

set_fixed_portfolio_tp_ratio(config, value)

set_trailing_fixed_portfolio_tp_ratio(config, value)

set_atr_k_portfolio_tp(config, value)

set_vol_sigma_portfolio_tp(config, value)

set_log_vol_sigma_portfolio_tp(config, value)

set_single_max_weight(config, value)

set_global_max_hold_pct(config, value)

set_rebalance_mode(config, value)

set_rebalance_cycle(config, value)

set_weight_change_threshold(config, value)

set_skip_suspended(config, value)

set_output_type(config, value)
}
\arguments{
\item{config}{A configuration list from `default_backtest_config()`.}

\item{value}{New value for the parameter.}
}
\value{
Modified configuration list.
}
\description{
Modify a single field in the backtest configuration list.
}
