% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_weight_constraints}
\alias{set_weight_constraints}
\title{Set weight constraints}
\usage{
set_weight_constraints(
  config,
  single_max_weight = 0.95,
  global_max_hold_pct = 1,
  min_weight = 1e-06
)
}
\arguments{
\item{config}{Configuration list.}

\item{single_max_weight}{Maximum weight per asset.}

\item{global_max_hold_pct}{Maximum portfolio exposure.}

\item{min_weight}{Minimum effective weight.}
}
\value{
Modified configuration list.
}
\description{
Set weight constraints
}
